Move Params implementation to exiv2app.cpp file
